requirements.txt
pytest>=7.0.0
pytest-playwright>=0.4.0
playwright>=1.40.0
allure-pytest>=2.13.0
allure-python-commons>=2.13.0
pytest-xdist>=3.5.0
pytest-rerunfailures>=13.0
pytest-cov>=4.1.0
pytest-html>=4.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
requests>=2.31.0
Faker>=20.0.0
cryptography>=41.0.0
